завершает д/з 12 и проводит реструктуризацию больших блоков
