+++ index.html
@@ -68,6 +68,7 @@
                   id="upload-file"
                   class="img-upload__input visually-hidden"
                   name="filename"
+                  accept="image/png, image/jpeg"
                   required
                 />
                 <label
