+++ js/main.js
@@ -2,9 +2,8 @@ import {getPhotos} from './data.js';
 import {renderThumbNails} from './thumbnails.js';
 import {addThunbnailsEventListeners} from './gallery.js';
 import {showAlert} from './messages.js';
-import {onUploadForm} from './upload-form.js';
-import {closeUploadForm} from './upload-form.js';
-import {setUserFormSubmit} from './upload-form.js';
+import {onUploadForm, closeUploadForm} from './upload-form.js';
+import {setUserFormSubmit} from './form-validation.js';
 import {showFilterContainer, getDiscussedPhotos,getRandomPhotos, setDefaultClick, setRandomClick, setDiscussedClick} from './gallery-filter.js';
 import {debounce} from './util.js';
