+++ js/util.js
@@ -1,47 +1,21 @@
-function getRandomNumber(numberFrom,numberTo){
-  if(numberFrom < 0 || numberTo < 0){
-    return NaN;
-  }
-
-  const lower = Math.ceil(Math.min(numberFrom, numberTo));
-  const upper = Math.round(Math.max(numberTo, numberTo));
-
-  return Math.round(Math.random() * (upper - lower) + lower);
-}
-
-const UniqueRandomNumber = [];
-
-function getUniqueRandomNumber(min,max){
-  const uniqueNumber = getRandomNumber(min,max);
-  if(UniqueRandomNumber.includes(uniqueNumber)){
-    return getUniqueRandomNumber(min,max);
-  }
-  else if(UniqueRandomNumber.length >= (max - min + 1)) {
-    return null;
-  }
-  UniqueRandomNumber.push(uniqueNumber);
-  return uniqueNumber;
-}
-
-
-function isCommentLengthValid(comment,maxLength){
+const isCommentLengthValid = (comment,maxLength) => {
   if(comment.length > maxLength) {
     return false;
   }
   return true;
-}
+};
 isCommentLengthValid('comment',140);
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
-function debounce (callback, timeoutDelay) {
+const debounce = (callback, timeoutDelay) => {
   let timeoutId;
 
   return (...rest) => {
     clearTimeout(timeoutId);
     timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
   };
-}
+};
 
 
-export {getRandomNumber, getUniqueRandomNumber,isCommentLengthValid, isEscapeKey, debounce};
+export {isCommentLengthValid, isEscapeKey, debounce};
